% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{determine_probs}
\alias{determine_probs}
\title{Determine a table of transition probabilities dependend on age}
\usage{
determine_probs(employment_data)
}
\description{
Determine a table of transition probabilities dependend on age
}
