% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_markov.R
\name{extend_incdec_table}
\alias{extend_incdec_table}
\title{Extends an inc/decrement table
by the  by the number individuals
staying in their state at each age.}
\usage{
extend_incdec_table(icdc_df)
}
\arguments{
\item{icdc_df}{\itemize{
\item inc/decrement table such as \link{occupation1977stlouis} dataset
}}
}
\value{
extended dataframe
}
\description{
Extends an inc/decrement table
by the  by the number individuals
staying in their state at each age.
}
