% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_markov.R
\name{max_l_softmax_estimate}
\alias{max_l_softmax_estimate}
\title{Estimates parameters via max likelihood
for a softmax model of the markov transition matrices
dependent on age.}
\usage{
max_l_softmax_estimate(incdec_df, age_range)
}
\arguments{
\item{incdec_df}{\itemize{
\item An inc/decrement table with such as produced by \code{\link[=extend_incdec_table]{extend_incdec_table()}}.
}}

\item{age_range}{\itemize{
\item The age range for which the model should be estimated.
}}
}
\value{
parameters for softmax model.
}
\description{
Estimates parameters via max likelihood
for a softmax model of the markov transition matrices
dependent on age.
}
