% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{occupation1977stlouis}
\alias{occupation1977stlouis}
\title{St. Louis Occupation data}
\format{
\subsection{\code{occupation1977stlouis}}{

A data frame with 60 rows for ages 16 until 75:
\describe{
\item{age}{Age x}
\item{total}{total individuals}
\item{active}{Occupied individuals at age x}
\item{inactive}{Not occupied individuals}
\item{inactive_to_active}{Number of inactive individuals added to current active population during x,x+1}
\item{active_to_inactive}{Number of active individuals added to current inactive population during x,x+1}
\item{active_to_death}{Number of active individuals which died during x,x+1.}
\item{inactive_to_death}{Number of inactive individuals which died during x,x+1.}
\item{death_total}{Total number of deaths}
}
}
}
\source{
Tables of Working Life: The Increment-Decrement Model U.S. Department of Labor November 1982 Bulletin 2135
}
\usage{
occupation1977stlouis
}
\description{
Documenting the occupation status and change a cohort of 100000 individuals in St. Louis from 1977.
}
\keyword{datasets}
