% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{occupation_insurance_one_time_premium}
\alias{occupation_insurance_one_time_premium}
\title{Computes the single premium paid in advance
for an occupation insurance, which pays
a fixed period at each policy anniversary,
if the insuree is not employed. Time units are typically years.}
\usage{
occupation_insurance_one_time_premium(
  x,
  duration,
  benefit,
  v,
  transition_probs,
  active = TRUE
)
}
\arguments{
\item{x}{\itemize{
\item Age of the insuree at policy inception.
}}

\item{duration}{\itemize{
\item Policy duration.
}}

\item{benefit}{\itemize{
\item Benefit paid at the end of each year.
}}

\item{v}{\itemize{
\item Constant discount factor
}}

\item{transition_probs}{\itemize{
\item A table of transition probabilities as produced by \code{\link[=determine_probs]{determine_probs()}}
}}

\item{active}{\itemize{
\item True if the insuree has an occupation at policy inception.
}}
}
\value{
Premium to be paid
}
\description{
Computes the single premium paid in advance
for an occupation insurance, which pays
a fixed period at each policy anniversary,
if the insuree is not employed. Time units are typically years.
}
