% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{occupation_insurance_recurrent_premium}
\alias{occupation_insurance_recurrent_premium}
\title{Computes the constant recurrent/natural premium paid in advance
for an occupation insurance, which pays
a fixed period at each policy anniversary,
if the insuree is not employed. Time units are typically years.}
\usage{
occupation_insurance_recurrent_premium(
  x,
  duration,
  benefit,
  v,
  transition_probs,
  active = TRUE,
  opt_premium_duration = NULL
)
}
\arguments{
\item{x}{\itemize{
\item Age of the insuree at policy inception.
}}

\item{duration}{\itemize{
\item Policy duration.
}}

\item{benefit}{\itemize{
\item Benefit paid at the end of each year.
}}

\item{v}{\itemize{
\item Constant discount factor
}}

\item{transition_probs}{\itemize{
\item A table of transition probabilities as produced by \code{\link[=determine_probs]{determine_probs()}}
}}

\item{active}{\itemize{
\item True if the insuree has an occupation at policy inception.
}}

\item{opt_premium_duration}{\itemize{
\item The no. of times the premium is paid, default is the policy duration.
}}
}
\value{
constant premium to be paid
}
\description{
Computes the constant recurrent/natural premium paid in advance
for an occupation insurance, which pays
a fixed period at each policy anniversary,
if the insuree is not employed. Time units are typically years.
}
