% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price.R
\name{occupation_insurance_recurrent_reserve}
\alias{occupation_insurance_recurrent_reserve}
\title{Computes the reserve (depends on state) at time \code{t}
for a policy as given by the parameters.}
\usage{
occupation_insurance_recurrent_reserve(
  t,
  x,
  duration,
  benefit,
  v,
  transition_probs,
  active = TRUE,
  opt_premium_duration = NULL
)
}
\arguments{
\item{t}{\itemize{
\item Time at which the reserve is considered.
}}

\item{x}{\itemize{
\item Age of the insuree at policy inception.
}}

\item{duration}{\itemize{
\item Policy duration.
}}

\item{benefit}{\itemize{
\item Benefit paid at the end of each year.
}}

\item{v}{\itemize{
\item Constant discount factor
}}

\item{transition_probs}{\itemize{
\item A table of transition probabilities as produced by \code{\link[=determine_probs]{determine_probs()}}
}}

\item{active}{\itemize{
\item True if the insuree has an occupation at policy inception.
}}

\item{opt_premium_duration}{\itemize{
\item The no. of periods the premium is paid, default is the policy duration.
}}
}
\value{
reserve at time \code{t}.
}
\description{
Computes the reserve (depends on state) at time \code{t}
for a policy as given by the parameters.
}
