% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{set_all_transition_matrices}
\alias{set_all_transition_matrices}
\title{For efficient caching, fix all one step transition matrices.}
\usage{
set_all_transition_matrices(x, transition_probs)
}
\description{
For efficient caching, fix all one step transition matrices.
}
