% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_markov.R
\name{transition_matrix_from_est}
\alias{transition_matrix_from_est}
\title{Creates a a transition matrix of age t
from a softmax model.}
\usage{
transition_matrix_from_est(est, x)
}
\arguments{
\item{est}{\itemize{
\item Parameters of softmax model such as \code{\link[=max_l_softmax_estimate]{max_l_softmax_estimate()}}
}}

\item{x}{\itemize{
\item ages for which a transition matrix should be created
}}
}
\description{
Creates a a transition matrix of age t
from a softmax model.
}
