% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{transition_matrix_x}
\alias{transition_matrix_x}
\title{Determine a transition probability matrix
for Markov Chain based calculations}
\usage{
transition_matrix_x(x, transition_probs)
}
\description{
Determine a transition probability matrix
for Markov Chain based calculations
}
