% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{transition_matrix_x_t}
\alias{transition_matrix_x_t}
\title{Obtain the transition matrix for an individual aged x
after t years.}
\usage{
transition_matrix_x_t(x, t, transition_probs)
}
\description{
Obtain the transition matrix for an individual aged x
after t years.
}
